+++ src/test/java/racingcar/InputValidatorTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.catchThrowable;
 
 import org.junit.jupiter.api.Test;
+import racingcar.validator.InputValidator;
 
 public class InputValidatorTest {
